+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to the "key-bindings-to-md" extension will be documented in this file.
 
+## [0.3.0]
+
+- Display all keys for 2 sets choice
+- Add any even sets number
+- Group keys by extension (will be displayed better in next versions)
+
 ## [0.2.0]
 
 - Add more info about the extension